+++ src/quantitative_analysis.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 // Linear regression function OLS
 std::tuple<string, string, string, string> linear_regression(
-    const std::unordered_map<std::string, std::vector<int>>& df,
+    const std::unordered_map<std::string, std::vector<size_t>>& df,
     const std::unordered_map<std::string, double>& quantitative_phenotype) {
 
     size_t num_samples = df.size();
